+++ gridopt/__init__.py
@@ -7,3 +7,4 @@
 #*****************************************************#
 
 from . import power_flow
+from . import scripts