+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 * Made test utils more flexible to be used from outside unittests.
 * Added HVDC, FACTS and vdep-loads to ACPF.
 * Added option for using linearized AC power flow equations in ACPF. 
+* Specified gridopt command-line utility using setup.py console entry point.
+* Added command-line utility option for writing network to file.
 
 Version 1.3.5
 -------------